[junu0516/airbnb#27] Chore: JsonHandler 별도 파일로 분리, JsonHandlable 프로토콜 추가
