+++ iOS/Airbnb/RoomDetail/RoomDetailRepository.swift
@@ -1,20 +1,6 @@
 import Foundation
 import OSLog
 
-struct JsonHandler {
-    private let encoder = JSONEncoder()
-    private let decoder = JSONDecoder()
-    
-    func convertJsonToObject<T: Decodable>(from data: Data, to targetType: T.Type) -> T? {
-        return try? decoder.decode(T.self, from: data)
-    }
-    
-    func convertObjectToJson<T: Encodable>(from object: T) -> Data? {
-        return try? encoder.encode(object)
-    }
-}
-
-
 struct RoomDetailRepository {
     private let networkHandler = NetworkHandler()
     private let jsonHandler = JsonHandler()
